requirements.txt
numpy
opencv-python
Pillow
yt-dlp
